+++ src/main/java/com/hubspot/jinjava/tree/ExpressionNode.java
@@ -21,7 +21,6 @@ import org.apache.commons.lang3.StringUtils;
 
 import com.hubspot.jinjava.interpret.JinjavaInterpreter;
 import com.hubspot.jinjava.lib.filter.EscapeFilter;
-import com.hubspot.jinjava.lib.tag.AutoEscapeTag;
 import com.hubspot.jinjava.tree.output.OutputNode;
 import com.hubspot.jinjava.tree.output.RenderedOutputNode;
 import com.hubspot.jinjava.tree.parse.ExpressionToken;
@@ -51,7 +50,7 @@ public class ExpressionNode extends Node {
       }
     }
 
-    if (interpreter.getContext().get(AutoEscapeTag.AUTOESCAPE_CONTEXT_VAR, Boolean.FALSE).equals(Boolean.TRUE)) {
+    if (interpreter.getContext().isAutoEscape()) {
       result = EscapeFilter.escapeHtmlEntities(result);
     }