+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Jinjava Releases #
 
+### Version 2.1.3 ([Maven Central](http://search.maven.org/#search%7Cga%7C1%7Cg%3A%22com.hubspot.jinjava%22%20AND%20v%3A%222.1.3%22)) ###
+
+* 
+
 ### Version 2.1.2 ([Maven Central](http://search.maven.org/#search%7Cga%7C1%7Cg%3A%22com.hubspot.jinjava%22%20AND%20v%3A%222.1.2%22)) ###
 
 * Use resolved path value in include tag cycle detection